+++ src/core/Zone.h
@@ -15,7 +15,7 @@ namespace instprof {
             EventItem e;
             e.tag.type = EventType::ZoneBegin;
             e.zoneBegin.time         = GetTime();
-            e.zoneBegin.callsiteInfo = reinterpret_cast<uint64_t>(info);
+            e.zoneBegin.callsiteInfo = reinterpret_cast<uintptr_t>(info);
             e.zoneBegin.threadID     = GetCurrentThreadID();
             Profiler::Get().EnqueueEvent(e);
         }
